# core/types: transaction and receipt encoding/decoding optimizations