--- a/core/types/transaction.go
+++ b/core/types/transaction.go
@@ -147,10 +147,16 @@ func (tx *Transaction) DecodeRLP(s *rlp.Stream) error {
 		return err
 	default:
 		// It's an EIP-2718 typed TX envelope.
-		var b []byte
-		if b, err = s.Bytes(); err != nil {
+		// First read the tx payload bytes into a temporary buffer.
+		b, buf, err := getPooledBuffer(size)
+		if err != nil {
 			return err
 		}
+		defer encodeBufferPool.Put(buf)
+		if err := s.ReadBytes(b); err != nil {
+			return err
+		}
+		// Now decode the inner transaction.
 		inner, err := tx.decodeTyped(b)
 		if err == nil {
 			tx.setDecoded(inner, uint64(len(b)))