… core/types/log.go