Just some minor optimizations I figured out a while ago. By using `ReadBytes` instead of
`Bytes` on the rlp stream, we can save the allocation of a temporary buffer for the typed tx
payload.

```
goos: linux
goarch: amd64
pkg: github.com/ethereum/go-ethereum/core/types
cpu: Intel(R) Core(TM) i5-3380M CPU @ 2.90GHz
                                │ types.old.txt │            types.new.txt            │
                                │    sec/op     │   sec/op     vs base                │
DecodeRLP/receipt-for-storage-2     1.080µ ± 4%   1.008µ ± 8%   -6.71% (p=0.007 n=20)
DecodeRLP/receipt-full-2            1.421µ ± 5%   1.333µ ± 3%   -6.19% (p=0.021 n=20)
DecodeRLP/access-transaction-2      3.584µ ± 9%   4.072µ ± 7%  +13.62% (p=0.001 n=20)
DecodeRLP/1559-transaction-2        4.444µ ± 7%   3.958µ ± 7%  -10.93% (p=0.002 n=20)

                                │ types.old.txt │            types.new.txt             │
                                │      B/s      │     B/s       vs base                │
DecodeRLP/receipt-for-storage-2   7.949Mi ±  4%   8.516Mi ± 7%   +7.14% (p=0.007 n=20)
DecodeRLP/receipt-full-2          181.2Mi ±  5%   193.1Mi ± 3%   +6.57% (p=0.021 n=20)
DecodeRLP/access-transaction-2    28.21Mi ± 10%   24.82Mi ± 7%  -12.00% (p=0.001 n=20)
DecodeRLP/1559-transaction-2      23.61Mi ±  8%   26.50Mi ± 6%  +12.26% (p=0.002 n=20)

                                │ types.old.txt │            types.new.txt             │
                                │     B/op      │    B/op     vs base                  │
DecodeRLP/receipt-for-storage-2      144.0 ± 0%   144.0 ± 0%        ~ (p=1.000 n=20)
DecodeRLP/receipt-full-2             393.0 ± 0%   393.0 ± 0%        ~ (p=1.000 n=20)
DecodeRLP/access-transaction-2       736.0 ± 0%   624.0 ± 0%  -15.22% (p=0.000 n=20)
DecodeRLP/1559-transaction-2         768.0 ± 0%   656.0 ± 0%  -14.58% (p=0.000 n=20)

                                │ types.old.txt │            types.new.txt            │
                                │   allocs/op   │ allocs/op   vs base                 │
DecodeRLP/receipt-for-storage-2      5.000 ± 0%   5.000 ± 0%       ~ (p=1.000 n=20)
DecodeRLP/receipt-full-2             4.000 ± 0%   4.000 ± 0%       ~ (p=1.000 n=20)
DecodeRLP/access-transaction-2       18.00 ± 0%   17.00 ± 0%  -5.56% (p=0.000 n=20)
DecodeRLP/1559-transaction-2         19.00 ± 0%   18.00 ± 0%  -5.26% (p=0.000 n=20)
```